Move Task from/to data functions to backend crate
