+++ crates/flutter_bridge/src/frb_generated.rs
@@ -29,7 +29,7 @@ use crate::api::error::*;
 use crate::api::repository::*;
 use crate::api::settings::*;
 use flutter_rust_bridge::for_generated::byteorder::{NativeEndian, ReadBytesExt, WriteBytesExt};
-use flutter_rust_bridge::for_generated::{transform_result_dco, Lifetimeable, Lockable};
+use flutter_rust_bridge::for_generated::{Lifetimeable, Lockable, transform_result_dco};
 use flutter_rust_bridge::{Handler, IntoIntoDart};
 use stride_core::event::*;
 use stride_plugin_manager::manifest::*;
@@ -42,7 +42,7 @@ flutter_rust_bridge::frb_generated_boilerplate!(
     default_rust_auto_opaque = RustAutoOpaqueMoi,
 );
 pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_VERSION: &str = "2.10.0";
-pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -739233487;
+pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -600290493;
 
 // Section: executor
 
@@ -2817,40 +2817,6 @@ fn wire__stride_core__task__task_default_impl(
         },
     )
 }
-fn wire__stride_core__task__task_from_data_impl(
-    port_: flutter_rust_bridge::for_generated::MessagePort,
-    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
-    rust_vec_len_: i32,
-    data_len_: i32,
-) {
-    FLUTTER_RUST_BRIDGE_HANDLER.wrap_normal::<flutter_rust_bridge::for_generated::SseCodec, _, _>(
-        flutter_rust_bridge::for_generated::TaskInfo {
-            debug_name: "task_from_data",
-            port: Some(port_),
-            mode: flutter_rust_bridge::for_generated::FfiCallMode::Normal,
-        },
-        move || {
-            let message = unsafe {
-                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
-                    ptr_,
-                    rust_vec_len_,
-                    data_len_,
-                )
-            };
-            let mut deserializer =
-                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
-            let api_input = <Vec<u8>>::sse_decode(&mut deserializer);
-            deserializer.end();
-            move |context| {
-                transform_result_sse::<_, ()>((move || {
-                    let output_ok =
-                        Result::<_, ()>::Ok(stride_core::task::Task::from_data(&api_input))?;
-                    Ok(output_ok)
-                })())
-            }
-        },
-    )
-}
 fn wire__stride_core__task__task_new_impl(
     ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
     rust_vec_len_: i32,
@@ -3009,40 +2975,6 @@ fn wire__stride_core__task__task_status_is_pending_impl(
                 transform_result_sse::<_, ()>((move || {
                     let output_ok =
                         Result::<_, ()>::Ok(stride_core::task::TaskStatus::is_pending(&api_that))?;
-                    Ok(output_ok)
-                })())
-            }
-        },
-    )
-}
-fn wire__stride_core__task__task_to_data_impl(
-    port_: flutter_rust_bridge::for_generated::MessagePort,
-    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
-    rust_vec_len_: i32,
-    data_len_: i32,
-) {
-    FLUTTER_RUST_BRIDGE_HANDLER.wrap_normal::<flutter_rust_bridge::for_generated::SseCodec, _, _>(
-        flutter_rust_bridge::for_generated::TaskInfo {
-            debug_name: "task_to_data",
-            port: Some(port_),
-            mode: flutter_rust_bridge::for_generated::FfiCallMode::Normal,
-        },
-        move || {
-            let message = unsafe {
-                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
-                    ptr_,
-                    rust_vec_len_,
-                    data_len_,
-                )
-            };
-            let mut deserializer =
-                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
-            let api_that = <stride_core::task::Task>::sse_decode(&mut deserializer);
-            deserializer.end();
-            move |context| {
-                transform_result_sse::<_, ()>((move || {
-                    let output_ok =
-                        Result::<_, ()>::Ok(stride_core::task::Task::to_data(&api_that))?;
                     Ok(output_ok)
                 })())
             }
@@ -4586,21 +4518,19 @@ fn pde_ffi_dispatcher_primary_impl(
         70 => wire__crate__api__settings__settings_save_impl(port, ptr, rust_vec_len, data_len),
         71 => wire__crate__api__settings__ssh_keys_impl(port, ptr, rust_vec_len, data_len),
         72 => wire__stride_core__task__task_default_impl(port, ptr, rust_vec_len, data_len),
-        73 => wire__stride_core__task__task_from_data_impl(port, ptr, rust_vec_len, data_len),
-        75 => wire__stride_core__task__task_priority_as_str_impl(port, ptr, rust_vec_len, data_len),
-        76 => {
+        74 => wire__stride_core__task__task_priority_as_str_impl(port, ptr, rust_vec_len, data_len),
+        75 => {
             wire__stride_core__task__task_priority_default_impl(port, ptr, rust_vec_len, data_len)
         }
-        77 => wire__stride_core__task__task_status_default_impl(port, ptr, rust_vec_len, data_len),
-        78 => {
+        76 => wire__stride_core__task__task_status_default_impl(port, ptr, rust_vec_len, data_len),
+        77 => {
             wire__stride_core__task__task_status_is_pending_impl(port, ptr, rust_vec_len, data_len)
         }
-        79 => wire__stride_core__task__task_to_data_impl(port, ptr, rust_vec_len, data_len),
-        81 => wire__stride_core__task__task_with_uuid_impl(port, ptr, rust_vec_len, data_len),
-        82 => wire__crate__api__plugin_manager__toggle_impl(port, ptr, rust_vec_len, data_len),
-        83 => wire__crate__api__logging__trace_impl(port, ptr, rust_vec_len, data_len),
-        84 => wire__stride_core__task__uda__uda_default_impl(port, ptr, rust_vec_len, data_len),
-        85 => wire__crate__api__logging__warn_impl(port, ptr, rust_vec_len, data_len),
+        79 => wire__stride_core__task__task_with_uuid_impl(port, ptr, rust_vec_len, data_len),
+        80 => wire__crate__api__plugin_manager__toggle_impl(port, ptr, rust_vec_len, data_len),
+        81 => wire__crate__api__logging__trace_impl(port, ptr, rust_vec_len, data_len),
+        82 => wire__stride_core__task__uda__uda_default_impl(port, ptr, rust_vec_len, data_len),
+        83 => wire__crate__api__logging__warn_impl(port, ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }
@@ -4667,8 +4597,8 @@ fn pde_ffi_dispatcher_sync_impl(
             data_len,
         ),
         69 => wire__crate__api__settings__settings_new_impl(ptr, rust_vec_len, data_len),
-        74 => wire__stride_core__task__task_new_impl(ptr, rust_vec_len, data_len),
-        80 => wire__stride_core__task__task_urgency_impl(ptr, rust_vec_len, data_len),
+        73 => wire__stride_core__task__task_new_impl(ptr, rust_vec_len, data_len),
+        78 => wire__stride_core__task__task_urgency_impl(ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }
@@ -6227,7 +6157,7 @@ mod io {
     use flutter_rust_bridge::for_generated::byteorder::{
         NativeEndian, ReadBytesExt, WriteBytesExt,
     };
-    use flutter_rust_bridge::for_generated::{transform_result_dco, Lifetimeable, Lockable};
+    use flutter_rust_bridge::for_generated::{Lifetimeable, Lockable, transform_result_dco};
     use flutter_rust_bridge::{Handler, IntoIntoDart};
     use stride_core::event::*;
     use stride_plugin_manager::manifest::*;
